+++ libraries/extractor/src/main/java/androidx/media3/extractor/mp4/Mp4Extractor.java
@@ -738,7 +738,9 @@ public final class Mp4Extractor implements Extractor, SeekMap {
           roleFlags |=
               firstVideoTrackIndex == C.INDEX_UNSET ? C.ROLE_FLAG_MAIN : C.ROLE_FLAG_ALTERNATE;
         }
-        if (trackDurationUs > 0 && trackSampleTable.sampleCount > 0) {
+        if (track.format.frameRate == Format.NO_VALUE
+            && trackDurationUs > 0
+            && trackSampleTable.sampleCount > 0) {
           float frameRate = trackSampleTable.sampleCount / (trackDurationUs / 1000000f);
           formatBuilder.setFrameRate(frameRate);
         }